back/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=0.19.0
elevenlabs>=0.2.24
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-multipart~=0.0.6
openai-whisper
sounddevice
httpx
ffmpeg-python